requirements.txt
pydantic~=2.9.2
fastapi~=0.115.4
loguru~=0.7.2
uvicorn~=0.32.0
pywebview~=5.3.2